// src/widgets/Menu/types.ts
import { Login } from "../WalletModal/types";

export interface LangType {
  code: string;
  language: string;
}

export interface Profile {
  username?: string;
  image?: string;
  profileLink: string;
  noProfileLink: string;
  showPip?: boolean;
}

export interface PushedProps {
  isPushed: boolean;
  pushNav: (isPushed: boolean) => void;
}

export interface NavTheme {
  background: string;
  hover: string;
}

export interface MenuSubEntry {
  label: string;
  href: string;
  icon: string;
  items?: MenuSubSubEntry[];
  calloutClass?: string;
  initialOpenState?: boolean;
  disabled?: boolean;
  balance?: number;
  cost?: number;
  coin?: string;
}

export interface MenuSubSubEntry {
  label: string;
  href: string;
  icon: string;
  calloutClass?: string;
  disabled?: boolean;
  balance?: number;
  cost?: number;
  coin?: string;
}

export interface MenuEntry {
  label: string;
  icon: string;
  items?: MenuSubEntry[];
  href?: string;
  calloutClass?: string;
  initialOpenState?: boolean;
}

export interface PanelProps {
  isDark: boolean;
  toggleTheme: (isDark: boolean) => void;
  cakePriceUsd?: number;
  currentLang: string;
  langs: LangType[];
  setLang: (lang: LangType) => void;
  links: Array<MenuEntry>;
  priceLink: string;
}

export interface NavProps extends PanelProps {
  account?: string;
  login: Login;
  profile?: Profile;
  logout: () => void;
  onBuyCryptoWithSimplex: () => void;
  showBalancePanel?: boolean;
  showBalanceContol?: boolean;
  totalCost?: number;
  showBuyButton?: boolean;
  showContractButton?: boolean;
  showHowButton?: boolean;
  onAddToken: () => void;
}